+++ examples/cpp/src/man_dataset.cpp
@@ -7,6 +7,7 @@
 #include <thread>
 #include <chrono>
 #include <filesystem>
+#include <iostream>
 
 namespace fs = std::filesystem;
 namespace man_ds = man::dataset;
@@ -23,34 +24,11 @@ int main(int argc, char** argv){
 
   // Print scenes and select one
   const auto& scene = select_scene(dataset.get_scene_manager());
+  std::cout << "Selected scene:\n\t" << scene << std::endl;
   const auto& sensor = select_sensor(dataset.get_sensor_manager());
-  
+  std::cout << "Selected sensor:\n\t" << sensor << std::endl;
 
-
-  
-  // const auto prime_sensor = sensor_manager[man_ds::Token(sensor_tokens[0])];
-  
-  // std::cout << "Sensor info:\n" << prime_sensor << std::endl;
-  // const auto data_item = prime_sensor[0];
-  // std::cout << "First data item:\n"
-  //           << "\tFilename: " << data_item.get_filename() << "\n"
-  //           << "\tTimestamp: " << data_item.get_timestamp() << std::endl;
-  
-  // if (prime_sensor.get_modality() == "lidar"){
-  //   pcl::visualization::PCLVisualizer viewer("Point Cloud Viewer");
-  //   auto pointcloud = std::make_shared<man_ds::data::PointCloudT>();
-  //   viewer.addPointCloud<man_ds::data::PointT>(pointcloud, prime_sensor.get_channel());
-
-  //   for (size_t i = 0; i < prime_sensor.size(); ++i){
-  //     const auto data_item = prime_sensor[i];
-  //     std::cout << "Data item " << i << ":\n"
-  //               << "\tFilename: " << data_item.get_filename() << "\n"
-  //               << "\tTimestamp: " << data_item.get_timestamp() << std::endl;
-  //     pointcloud = man_ds::data::read_pointcloud(data_item.get_filename());
-  //     std::cout << "Point cloud loaded with " << pointcloud->size() << " points." << std::endl;
-  //     viewer.updatePointCloud<man_ds::data::PointT>(pointcloud, prime_sensor.get_channel());
-  //     viewer.spinOnce (1000);
-  //   }
-  // }
+  const auto sensor_data = dataset.get_data(scene, sensor);
+  std::cout << "Data count:\n" << sensor_data.size() << std::endl;
   return 0;
 }